[core] Breaks down rake tasks
While documenting our release process, we can see some things that don't
need to be locked together are. Mostly generating files and uploading
them or packaging the release and immediately pushing.

Breaking these down will allow refactoring the steps or regrouping them
so preparing a release is not coupled with pushing the new release and
requiring lots of credentials (rubygems, github etc.).