Stops running of 'docs' rake task in release
The `release` task has been running the `docs` task as part of itself.
This happens after the release has been tagged. However it currently
just generates and uploads the fog.io site which are ignored anyway.

So preparation of the gem has been linked to updating the website with
minimal benefit.

We have decided to move the website to github pages so the existing
process is no much use and will be changing.